Refactoring: Substituir método p/ met. objeto Mau-cheiro identificado: classe inchada Descrição da refatoração: refatorado os métodos existe_argumento_em_branco e valida_argumento_positivos para uma nova classe genérica que é herdada pela classe Estacionamento aumentando a coesão das classes e diminuindo o tamanho da classe Estacionamento. Classes/métodos/atributos afetados: * Métodos foram extraídos para a class ObjetoGenérico que é herdada por Estacionamento
